refactorizacion y organizacion de las clases
