+++ coworking/coworking/src/main/java/com/backend/digitalhouse/coworking/dto/security/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.backend.digitalhouse.coworking.dto;
+package com.backend.digitalhouse.coworking.dto.security;
 
 public class AuthenticationResponse {
   private String jwt;
