+++ coworking/coworking/src/main/java/com/backend/digitalhouse/coworking/entity/Imagen.java
@@ -21,7 +21,6 @@ public class Imagen {
     @JoinColumn(name="salas_id_sala")
     private Long idSala;*/
 
-
     public Imagen() {
     }
