+++ coworking/coworking/src/main/java/com/backend/digitalhouse/coworking/entity/ReservaEspacio.java
@@ -1,7 +1,6 @@
 package com.backend.digitalhouse.coworking.entity;
 
 import jakarta.persistence.*;
-
 import java.time.LocalDateTime;
 
 @Entity
